% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lending}
\alias{lending}
\title{Lending data set}
\format{
A data frame with 10,000 rows and 27 variables; each row represents
one accepted loan application from the Lending Club platform. The 'default' variable indicates
whether the applicant defaulted on the loan (0 = did not default; 1 = defaulted).
For definitions of the remaining predictor variables (listed below), please refer to the data dictionary
provided in the Lending Club case.
\describe{
  \item{id}{A unique identifier for the dataset}
  \item{default}{A binary variable indicating whether the applicant defaulted on the loan (1) or not (0)}
  \item{home_ownership}{}
  \item{annual_inc}{}
  \item{mort_acc}{}
  \item{purpose}{}
  \item{application_type}{}
  \item{loan_amnt}{}
  \item{installment}{}
  \item{int_rate}{}
  \item{loan_term}{}
  \item{verification_status}{}
  \item{dti}{}
  \item{fico_range_low}{}
  \item{fico_range_high}{}
  \item{total_acc}{}
  \item{total_bal_ex_mort}{}
  \item{avg_cur_bal}{}
  \item{total_bc_limit}{}
  \item{open_acc}{}
  \item{inq_last_6mths}{}
  \item{tot_coll_amt}{}
  \item{delinq_amnt}{}
  \item{age_earliest_cred_mo}{}
  \item{total_rec_prncp}{}
  \item{pymnt_plan}{}
  \item{debt_settlement_flag}{}
  }
}
\usage{
lending
}
\description{
A dataset containing simulated data for a fictional company that is studying the
factors that are associated with employee attrition (choosing to leave the firm).
}
\examples{
data(lending)
}
\keyword{datasets}
