% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{purchases}
\alias{purchases}
\title{Purchases data set}
\format{
A data frame with 500 rows and 4 variables; each row represents
one customer (simulated, for a fictional customer) with their demographic information
and an indicator for whether they have made a purchase from the company.
\describe{
  \item{age}{Customer age in years}
  \item{gender}{Customer gender; M = man, W = woman, X = Non-binary or other not listed}
  \item{income}{Customer annual income in USD}
  \item{purchase}{Binary variable indicating whether the customer made a purchase; 1 = yes, 0 = no}
  }
}
\usage{
purchases
}
\description{
A dataset containing simulated customer data representing purchase status (1 = made a purchase, 0 = did not
make a purchase) and selected customer demographics (age, gender, and annual income).
}
\examples{
data(purchases)
}
\keyword{datasets}
