% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{purchases50}
\alias{purchases50}
\title{Purchases50 data set}
\format{
A data frame with 50 rows and 3 variables; each row represents
one customer (simulated, for a fictional customer) with their demographic information
and an indicator for whether they have made a purchase from the company.
\describe{
  \item{age}{Customer age in years}
  \item{income}{Customer annual income in USD}
  \item{purchase}{Binary variable indicating whether the customer made a purchase; 1 = yes, 0 = no}
  }
}
\usage{
purchases50
}
\description{
A dataset containing simulated customer data representing purchase status (1 = made a purchase, 0 = did not
make a purchase) and selected customer demographics (age and annual income).
This dataset is a subset of 'purchases' containing just 50 observations and two demographic variables.
}
\examples{
data(purchases50)
}
\keyword{datasets}
